Skeleton API (#4)
